refactor: remove conversationManager facade and update imports
- Delete src/conversationManager.js facade file
- Update all production code imports to use core/conversation directly:
  - handlers: referenceHandler, messageHandler, personalityHandler, dmHandler
  - commands: activate, deactivate, autorespond, status, reset
- All functionality preserved, just removing the indirection layer
- Tests will be updated in the next commit

Part of facade removal effort to simplify the codebase
