+++ src/handlers/personalityHandler.js
@@ -16,7 +16,7 @@ const webhookUserTracker = require('../utils/webhookUserTracker');
 const referenceHandler = require('./referenceHandler');
 const { detectMedia } = require('../utils/media');
 const { MARKERS } = require('../constants');
-const { recordConversation, isAutoResponseEnabled } = require('../conversationManager');
+const { recordConversation, isAutoResponseEnabled } = require('../core/conversation');
 const requestTracker = require('../utils/requestTracker');
 const personalityAuth = require('../utils/personalityAuth');
 const threadHandler = require('../utils/threadHandler');
